requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
Pillow>=9.0.0
scipy>=1.8.0
color-matcher
matplotlib
huggingface_hub
mss
opencv-python
peft